Modulation is done
