+++ js/main.js
@@ -1,103 +1,4 @@
-// Количество постов
-const SIMILAR_POST_COUNT = 25;
-// Пул комментариев
-const COMMENT_MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-// Пул имён комментаторов
-const COMMENTATORS_NAMES = [
-  'Андрей', 'Аня', 'Даниил', 'Полина', 'Алина', 'Константин', 'Арсений', 'Наташа',
-  'Матвей', 'Ника', 'Юра', 'Ильяс', 'Дмитрий', 'Влад', 'Алексей', 'Серёжа', 'Ваня',
-  'Макар', 'Андрей'
-];
-// Пул описаний постов в Кекстаграме
-const POSTS_DESCRIPTIONS = [
-  'Меня трудно найти, легко потерять и невозможно забыть',
-  'Настоящий мужчина, как волк: либо один, либо с одной волчицей навсегда! А бегать за овцами — это удел баранов!',
-  'Я не конфетка, чтобы быть всем по вкусу. Я орешек, который не всем по зубам!',
-  'Бессмысленно осмысливать смысл неосмысленными мыслями',
-  'Самая большая ненависть возникает к тем, кто сумел дотронуться к сердцу, а затем плюнул в душу',
-  'Кому-то любимая… Кому-то дочь… Кому-то сестра… Кому-то подруга… А кому-то поперёк горла!!! Горжусь своей универсальностью!!!',
-  'Сегодня меня никто не поздравил с днем рождения. Ничего удивительного, он у меня не сегодня',
-  'Я никого никогда не держу… Потому что тот, кто любит — все равно останется, а тот, кто не любит — все равно уйдет',
-  'Не люблю, когда меня с кем-то сравнивают. Не устраиваю? Отдыхай!',
-  'Я не ангел и не демон. И совсем не идеал. Ну, такой я получился. Так Господь меня создал!',
-  'О чем мечтаю я?! Да просто жить… дышать… любить… и знать, что я любима! И каждое мгновение ценить… Поскольку наша жизнь неповторима',
-  'Мы все как карандаши. Каждый рисует свою судьбу. Кто-то ломается, кто-то тупит, а кто-то затачивается и движется вперед',
-  'Я не нарушаю правила, я просто живу по своим',
-  'Живу для тех, кому нужен… дружу лишь с теми, в ком уверен… общаюсь с теми, кто приятен… И благодарен тем, кто ценит',
-  'Все могут видеть, как ты выглядишь внешне, но очень мало тех, кто знает, что у тебя в душе',
-  'В жизни есть конец всему. Любви… Друзьям… Страданьям… Но нет конца лишь одному — Воспоминаниям…',
-  'Не стоит слишком настойчиво интересоваться моей жизнью. Она может оказаться настолько интересной, что вы разочаруетесь в своей',
-  'Вы ошибаетесь, если говорите, что знаете меня. Каждый знает меня настолько, насколько я позволяю себя узнать',
-  'У меня пушистый характер! Да, да! Как у ежика!',
-  'В жизни, как под дождем — наступает момент, когда уже просто все равно…',
-  'Не знакомлюсь, не даю телефончик, не соглашаюсь на свидания. Но… уважаю ваш выбор =)',
-  'ЖИВУ, ЦВЕТУ и ХОРОШЕЮ! О том, что было — не жалею! О том, что будет — не гадаю! А то, что есть — я обожаю!!!'
-];
+import './util';
+import {createPosts} from "./data";
 
-// Функция, возвращающая случайное целое число из переданного диапазона включительно
-const getRandomIntInclusive = (a, b) => {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-}
-
-// Функция для проверки длины введённого комментария
-const isCorrectLength = (comment, maxLength) => {
-  return comment.length <= maxLength;
-}
-
-// Функция для получения рандомного элемента из массива elements
-const getRandomArrayElement = (elements) => {
-  return elements[getRandomIntInclusive(0, elements.length - 1)];
-};
-
-// Функция, генерирующая пост
-const createPost = (index) => {
-  // Массив с комментариями
-  let comments = [];
-  for (let i = 0; i <= getRandomIntInclusive(1, 5); i++) {
-    comments[i] = createComment(i);
-  }
-
-  return {
-    // Идентификатор описания (число от 1 до 25)
-    id: index,
-    // Строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25
-    url: `photos/${index}.jpg`,
-    // Описание фотографии (придумывается самостоятельно)
-    description: getRandomArrayElement(POSTS_DESCRIPTIONS),
-    // Число — количество лайков, поставленных фотографии (случайное число от 15 до 200)
-    likes: getRandomIntInclusive(15, 200),
-    // Массив объектов — список комментариев, оставленных другими пользователями к этой фотографии
-    comments: comments
-  };
-}
-
-// Функция, генерирующая комментарии
-const createComment = (index) => {
-  return {
-    // Идентификатор
-    id: index,
-    // строка, значение которой формируется по правилу img/avatar-{{случайное число от 1 до 6}}.svg
-    avatar: `img/avatar-${getRandomIntInclusive(1, 6)}.svg`,
-    // Текст комментария
-    message: getRandomArrayElement(COMMENT_MESSAGES),
-    // Набор имён для комментаторов
-    name: getRandomArrayElement(COMMENTATORS_NAMES)
-  };
-}
-
-// Массив с временными данными
-const similarPosts = [];
-for (let i = 1; i <= SIMILAR_POST_COUNT; i++) {
-  similarPosts[i] = createPost(i);
-}
+console.log(createPosts());
